Move LuDown JS files from /bin/ to /lib/ fix #276
