+++ Ludown/lib/ludown-parse.js
@@ -5,7 +5,7 @@
  */
 const program = require('commander');
 const chalk = require('chalk');
-const retCode = require('../lib/enums/CLI-errors');
+const retCode = require('./enums/CLI-errors');
 program.Command.prototype.unknownOption = function (flag) {
     process.stderr.write(chalk.default.redBright(`\n  Unknown arguments: ${process.argv.slice(2).join(' ')}\n`));
     program.help();
