+++ Ludown/lib/ludown-translate.js
@@ -5,8 +5,8 @@
  */
 const program = require('commander');
 const chalk = require('chalk');
-const translate = require('../lib/translate');
-const retCode = require('../lib/enums/CLI-errors');
+const translate = require('../translate');
+const retCode = require('./enums/CLI-errors');
 program.Command.prototype.unknownOption = function (flag) {
     process.stderr.write(chalk.default.redBright(`\n  Unknown arguments: ${process.argv.slice(2).join(' ')}\n`));
     program.help();
