+++ Ludown/package.json
@@ -2,12 +2,12 @@
   "name": "ludown",
   "version": "1.0.31",
   "description": "Command line tool to convert .lu files to LUIS JSON and QnA maker JSON files",
-  "main": "lib\\exports.js",
+  "main": "lib/exports.js",
   "scripts": {
     "test": "nyc --reporter=html mocha --timeout 100000"
   },
   "bin": {
-    "ludown": "./bin/ludown.js"
+    "ludown": "./bin/ludown"
   },
   "repository": {
     "type": "git",
