+++ Ludown/test/All-examples-pass.js
@@ -6,7 +6,7 @@ var chai = require('chai');
 var assert = chai.assert;
 var path = require('path');
 const {exec} = require('child_process');
-const ludown = path.resolve('./bin/ludown.js');
+const ludown = path.resolve('./bin/ludown');
 const txtfile = require('read-text-file');
 const pathToOutputFolder = path.resolve('./test/output');
 const fs = require('fs');
