+++ Ludown/test/translate.js
@@ -7,7 +7,7 @@ const assert = chai.assert;
 const translate = require('../lib/translate');
 const path = require('path');
 const {exec} = require('child_process');
-const ludown = path.resolve('./bin/ludown.js');
+const ludown = path.resolve('./bin/ludown');
 const txtfile = require('read-text-file');
 const trHelpers = require('../lib/translate-helpers');
 const pathToOutputFolder = path.resolve('./test/output');
